Move beam energy to be a float to avoid conversion

+++ Ecal/include/Ecal/EcalVetoProcessor.h
@@ -181,8 +181,8 @@ class EcalVetoProcessor : public framework::Producer {
 
   double bdtCutVal_{0};
 
-  double beamEnergyMeV_{0};
-  double linreg_radius_{0};
+  float beamEnergyMeV_{0};
+  float linreg_radius_{0};
 
   bool verbose_{false};
